+++ report/stage_2/REPORT.md
@@ -90,6 +90,8 @@
 
 ### Оптимизация метода `range`
 
+[3e264c42b96235374f0b2b4df0bf62510a2a81f1]()
+
 Сразу стоит оговориться про написанный тест `hugeRecordsWriteReadFlushTest`.
 Если просто убрать блок `synchronized` у метода `range`, все тесты будут пройдены.
 Однако разработанный тест завалится, что есть хорошо. Потому что на самом деле
