+++ report/stage_2/profiling/profiler_alloc_getput.html
@@ -314,17 +314,17 @@ f(9,37517,801,0,'java/util/concurrent/ConcurrentSkipListMap$SubMap.values')
 f(10,37517,801,1,'java.util.concurrent.ConcurrentSkipListMap$Values')
 f(9,38318,1625,0,'java/util/concurrent/ConcurrentSkipListMap$Values.iterator')
 f(10,38318,1625,1,'java.util.concurrent.ConcurrentSkipListMap$SubMap$SubMapValueIterator')
-f(9,39943,4767,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(9,39943,4767,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(9,44710,1993,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.filterTombstones')
-f(10,44710,799,1,'ru.mail.polis.lsm.artem_drozdov.TombstonesFilterIterator')
+f(10,44710,799,1,'ru.mail.polis.lsm.artem_drozdov.iterators.TombstonesFilterIterator')
 f(10,45509,1194,0,'ru/mail/polis/lsm/artem_drozdov/TombstonesFilterIterator.<init>')
-f(11,45509,1194,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(11,45509,1194,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(9,46703,2362,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.map')
 f(10,46703,2362,0,'java/util/concurrent/ConcurrentSkipListMap.subMap')
 f(11,46703,2362,0,'java/util/concurrent/ConcurrentSkipListMap.subMap')
 f(12,46703,2362,1,'java.util.concurrent.ConcurrentSkipListMap$SubMap')
 f(9,49065,2343,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(10,49065,2343,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(10,49065,2343,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(9,51408,741642,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.sstableRanges')
 f(10,51408,1175,1,'java.util.ArrayList')
 f(10,52583,15582,0,'java/util/ArrayList.<init>')
@@ -333,49 +333,49 @@ f(11,68080,85,2,'java.lang.Object[]')
 f(10,68165,393865,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(11,68165,3167,0,'java/util/ArrayList.subList')
 f(12,68165,3167,1,'java.util.ArrayList$SubList')
-f(11,71332,2349,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(11,71332,2349,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(11,73681,387210,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(12,73681,6301,0,'java/util/ArrayList$SubList.subList')
 f(13,73681,6301,1,'java.util.ArrayList$SubList')
-f(12,79982,4562,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(12,79982,4562,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(12,84544,373976,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(13,84544,12405,0,'java/util/ArrayList$SubList.subList')
 f(14,84544,12405,1,'java.util.ArrayList$SubList')
-f(13,96949,9338,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(13,96949,9338,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(13,106287,347611,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(14,106287,24911,0,'java/util/ArrayList$SubList.subList')
 f(15,106287,24911,1,'java.util.ArrayList$SubList')
-f(14,131198,18711,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(14,131198,18711,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(14,149909,294559,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(15,149909,49770,0,'java/util/ArrayList$SubList.subList')
 f(16,149909,49770,1,'java.util.ArrayList$SubList')
-f(15,199679,37604,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(15,199679,37604,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(15,237283,188449,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(16,237283,35828,0,'java/util/ArrayList$SubList.subList')
 f(17,237283,35828,1,'java.util.ArrayList$SubList')
-f(16,273111,74599,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(16,273111,74599,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(16,347710,40548,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
-f(17,347710,27231,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(17,347710,27231,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(17,374941,13317,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(18,374941,13317,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(18,374941,13317,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(16,388258,37474,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(17,388258,37474,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(17,388258,37474,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(15,425732,18736,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(16,425732,18736,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(16,425732,18736,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(14,444468,9430,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(15,444468,9430,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(15,444468,9430,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(13,453898,4622,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(14,453898,4622,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(14,453898,4622,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(12,458520,2371,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(13,458520,2371,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(13,458520,2371,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(11,460891,1139,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(12,460891,1139,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(12,460891,1139,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(10,462030,331020,0,'ru/mail/polis/lsm/artem_drozdov/SSTable.range')
 f(11,462030,231015,0,'java/nio/DirectByteBufferR.asReadOnlyBuffer')
 f(12,462030,231015,0,'java/nio/DirectByteBufferR.duplicate')
 f(13,462030,230909,1,'java.nio.DirectByteBufferR')
 f(13,692939,106,2,'java.nio.DirectByteBufferR')
-f(11,693045,100005,1,'ru.mail.polis.lsm.artem_drozdov.RangeIterator')
+f(11,693045,100005,1,'ru.mail.polis.lsm.artem_drozdov.iterators.RangeIterator')
 f(7,793050,85692,0,'ru/mail/polis/lsm/artem_drozdov/TombstonesFilterIterator.hasNext')
 f(8,793050,85692,0,'ru/mail/polis/lsm/artem_drozdov/PeekingIterator.peek')
 f(9,793050,85692,0,'ru/mail/polis/lsm/artem_drozdov/MergeIterator.next')
