+++ report/stage_2/profiling/profiler_alloc_range_get.html
@@ -274,17 +274,17 @@ f(9,32932,855,0,'java/util/concurrent/ConcurrentSkipListMap$SubMap.values')
 f(10,32932,855,1,'java.util.concurrent.ConcurrentSkipListMap$Values')
 f(9,33787,1638,0,'java/util/concurrent/ConcurrentSkipListMap$Values.iterator')
 f(10,33787,1638,1,'java.util.concurrent.ConcurrentSkipListMap$SubMap$SubMapValueIterator')
-f(9,35425,4783,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(9,35425,4783,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(9,40208,2006,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.filterTombstones')
-f(10,40208,787,1,'ru.mail.polis.lsm.artem_drozdov.TombstonesFilterIterator')
+f(10,40208,787,1,'ru.mail.polis.lsm.artem_drozdov.iterators.TombstonesFilterIterator')
 f(10,40995,1219,0,'ru/mail/polis/lsm/artem_drozdov/TombstonesFilterIterator.<init>')
-f(11,40995,1219,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(11,40995,1219,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(9,42214,2321,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.map')
 f(10,42214,2321,0,'java/util/concurrent/ConcurrentSkipListMap.subMap')
 f(11,42214,2321,0,'java/util/concurrent/ConcurrentSkipListMap.subMap')
 f(12,42214,2321,1,'java.util.concurrent.ConcurrentSkipListMap$SubMap')
 f(9,44535,2425,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(10,44535,2425,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(10,44535,2425,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(9,46960,901319,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.sstableRanges')
 f(10,46960,1198,1,'java.util.ArrayList')
 f(10,48158,18811,0,'java/util/ArrayList.<init>')
@@ -292,48 +292,48 @@ f(11,48158,18811,1,'java.lang.Object[]')
 f(10,66969,489593,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(11,66969,3087,0,'java/util/ArrayList.subList')
 f(12,66969,3087,1,'java.util.ArrayList$SubList')
-f(11,70056,2286,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(11,70056,2286,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(11,72342,483066,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(12,72342,6455,0,'java/util/ArrayList$SubList.subList')
 f(13,72342,6455,1,'java.util.ArrayList$SubList')
-f(12,78797,4808,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(12,78797,4808,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(12,83605,469400,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(13,83605,12656,0,'java/util/ArrayList$SubList.subList')
 f(14,83605,12656,1,'java.util.ArrayList$SubList')
-f(13,96261,9461,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(13,96261,9461,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(13,105722,442820,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(14,105722,25210,0,'java/util/ArrayList$SubList.subList')
 f(15,105722,25210,1,'java.util.ArrayList$SubList')
-f(14,130932,19182,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(14,130932,19182,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(14,150114,388994,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(15,150114,50406,0,'java/util/ArrayList$SubList.subList')
 f(16,150114,50406,1,'java.util.ArrayList$SubList')
-f(15,200520,37330,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(15,200520,37330,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(15,237850,282127,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
 f(16,237850,78921,0,'java/util/ArrayList$SubList.subList')
 f(17,237850,78921,1,'java.util.ArrayList$SubList')
-f(16,316771,75976,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(16,316771,75976,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(16,392747,88825,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.merge')
-f(17,392747,59198,1,'ru.mail.polis.lsm.artem_drozdov.PeekingIterator')
+f(17,392747,59198,1,'ru.mail.polis.lsm.artem_drozdov.iterators.PeekingIterator')
 f(17,451945,29627,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(18,451945,29627,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(18,451945,29627,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(16,481572,38405,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(17,481572,38405,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(17,481572,38405,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(15,519977,19131,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(16,519977,19131,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(16,519977,19131,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(14,539108,9434,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(15,539108,9434,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(15,539108,9434,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(13,548542,4463,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(14,548542,4463,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(14,548542,4463,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(12,553005,2403,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(13,553005,2403,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(13,553005,2403,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(11,555408,1154,0,'ru/mail/polis/lsm/artem_drozdov/LsmDAO.mergeTwo')
-f(12,555408,1154,1,'ru.mail.polis.lsm.artem_drozdov.MergeIterator')
+f(12,555408,1154,1,'ru.mail.polis.lsm.artem_drozdov.iterators.MergeIterator')
 f(10,556562,391717,0,'ru/mail/polis/lsm/artem_drozdov/SSTable.range')
 f(11,556562,273531,0,'java/nio/DirectByteBufferR.asReadOnlyBuffer')
 f(12,556562,273531,0,'java/nio/DirectByteBufferR.duplicate')
 f(13,556562,273531,1,'java.nio.DirectByteBufferR')
-f(11,830093,118186,1,'ru.mail.polis.lsm.artem_drozdov.RangeIterator')
+f(11,830093,118186,1,'ru.mail.polis.lsm.artem_drozdov.iterators.RangeIterator')
 f(7,948279,73612,0,'ru/mail/polis/lsm/artem_drozdov/TombstonesFilterIterator.hasNext')
 f(8,948279,73612,0,'ru/mail/polis/lsm/artem_drozdov/PeekingIterator.peek')
 f(9,948279,73612,0,'ru/mail/polis/lsm/artem_drozdov/MergeIterator.next')
